backend/requirements.txt
fastapi==0.115.8
uvicorn==0.30.6
pydantic==2.8.2
langchain==0.3.7
langgraph==0.2.40
langchain-google-genai==2.0.7
python-dotenv==1.0.1
